// apps/payment-service/src/infrastructure/services/paypal/paypal.type.ts
export type Intent = 'CAPTURE' | 'AUTHORIZE';
// ...
export interface Item {
  name: string;
  quantity: number;
  unit_amount: UnitAmount;
  url?: string;
  image_url?: string;
  tax?: UnitAmount;
}
// ...
export interface Amount {
  currency_code: string;
  value: string;
}

export interface Payee {
  email_address: string;
  merchant_id?: string;
}

export interface Phone {
  /**
   * string
   *
   * The phone type.
   */
  phone_type?: 'FAX' | 'HOME' | 'MOBILE' | 'OTHER' | 'PAGER' | undefined;
  /**
   * object
   *
   * The phone number, in its canonical international E.164 numbering plan format. Supports only the national_number property.
   *
   * required
   */
  phone_number: {
    /**
     * string
     *
     * The national number, in its canonical international E.164 numbering plan format. The combined length of the country calling code (CC) and the national number must not be greater than 15 digits.
     * The national number consists of a national destination code (NDC) and subscriber number (SN).
     *
     * required
     */
    national_number: string;
  };
}

export interface Name {
  /**
   * string
   *
   * When the party is a person, the party's given, or first, name.
   */
  given_name?: string | undefined;
  /**
   * string
   *
   * When the party is a person, the party's surname or family name. Also known as the last name. Required when the party is a person.
   * Use also to store multiple surnames including the matronymic, or mother's, surname.
   */
  surname?: string | undefined;
}

export interface TaxInfo {
  /**
   * string
   *
   * The customer's tax ID value
   *
   * required
   */
  tax_id: string;
  /**
   * string
   *
   * The customer's tax ID type.
   *
   * required
   */
  tax_id_type: TaxIdType;
}

export interface Payer {
  /**
   * The email address of the payer.
   */
  email_address?: string | undefined;
  /**
   * object
   *
   * The name of the payer. Supports only the given_name and surname properties.
   */
  name?: Name | undefined;

  /**
   * object
   *
   * The phone number of the customer. Available only when you enable the Contact Telephone Number option in the Profile & Settings for the merchant's PayPal account.
   * The phone.phone_number supports only national_number.
   */
  phone?: Phone | undefined;
  /**
   * string
   *
   * The birth date of the payer in YYYY-MM-DD format.
   */
  birth_date?: string | undefined;
  /**
   * object
   *
   * The tax information of the payer.
   */
  tax_info?: TaxInfo;
  /**
   * object
   *
   * The address of the payer. Supports only the address_line_1, address_line_2, admin_area_1, admin_area_2, postal_code, and country_code properties. Also referred to as the billing address of the customer.
   */
  address?: Address | undefined;
}
// ...
export interface Shipping {
  type?: Shippingype | undefined;
  address: Address;
  option: Array<{
    id: string;
    label: string;
    selected: string;
    type?: 'SHIPPING' | 'PICKUP' | 'PICKUP_IN_STORE' | 'PICKUP_FROM_PERSON';
    amount?: Amount;
  }>;
  name?: {
    full_name?: string;
  };
}

export interface Breakdown {
  /**
   * object
   *
   * The subtotal for all items. Required if the request includes purchase_units[].items[].unit_amount. Must equal the sum of (items[].unit_amount * items[].quantity) for all items.
   * item_total.value can not be a negative number.
   */
  item_total?: Amount;
  /**
   * object
   *
   * The shipping fee for all items within a given purchase_unit. shipping.value can not be a negative number.
   */
  shipping?: Amount;
  /**
   * object
   *
   * The handling fee for all items within a given purchase_unit. handling.value can not be a negative number.
   */
  handling?: Amount;
  /**
   * object
   *
   * The total tax for all items. Required if the request includes purchase_units.items.tax. Must equal the sum of (items[].tax * items[].quantity) for all items. tax_total.value can not be a negative number.
   */
  tax_total?: Amount;
  /**
   * object
   *
   * The insurance fee for all items within a given purchase_unit. insurance.value can not be a negative number.
   */
  insurance?: Amount;
  /**
   * object
   *
   * The shipping discount for all items within a given purchase_unit. shipping_discount.value can not be a negative number.
   */
  shipping_discount?: Amount;
  /**
   * object
   *
   * The discount for all items within a given purchase_unit. discount.value can not be a negative number.
   */
  discount?: Amount;
}

export interface PurchaseUnit {
  reference_id?: string;
  description?: string;
  custom_id?: string;
  invoice_id?: string;
  soft_descriptor?: string;
  items?: Array<Item>;
  amount: Amount & { breakdown?: Breakdown };
  payee?: Payee;
  payment_instruction?: object | any;
  shipping?: Shipping;
  supplementary_data?: object;
}
// ...
export interface ApplicationContext {
  /**
   * string
   *
   * DEPRECATED. The label that overrides the business name in the PayPal account on the PayPal site. The fields in application_context are now available in the experience_context object
   * under the payment_source which supports them (eg. payment_source.paypal.experience_context.brand_name). Please specify this field in the experience_context object instead of the application_context object.
   */
  brand_name?: string | undefined;
  /**
   * string
   *
   * DEPRECATED. The BCP 47-formatted locale of pages that the PayPal payment experience shows. PayPal supports a five-character code.
   */
  locale?: string;
  /**
   * string
   *
   * Default: "NO_PREFERENCE"
   *
   * DEPRECATED. DEPRECATED. The type of landing page to show on the PayPal site for customer checkout. The fields in application_context are now available in the experience_context object under
   * the payment_source which supports them (eg. payment_source.paypal.experience_context.landing_page). Please specify this field in the experience_context object instead of the application_context object.
   *
   */
  landing_page?: LandingPage;
  /**
   * string
   *
   * Default: "GET_FROM_FILE"
   *
   * DEPRECATED. DEPRECATED. The shipping preference:
   *
   * * Displays the shipping address to the customer.
   *
   * * Enables the customer to choose an address on the PayPal site.
   *
   * * Restricts the customer from changing the address during the payment-approval process.
   *
   */
  shipping_preference?: 'GET_FROM_FILE' | 'NO_SHIPPING' | 'SET_PROVIDED_ADDRESS';
  /**
   * string
   *
   * Default: "CONTINUE"
   *
   * DEPRECATED. Configures a Continue or Pay Now checkout flow. The fields in application_context are now available in the experience_context object under the payment_source which supports them
   * (eg. payment_source.paypal.experience_context.user_action). Please specify this field in the experience_context object instead of the application_context object.
   */
  user_action?: UserAction;
  /**
   * string
   *
   * DEPRECATED. The URL where the customer is redirected after the customer approves the payment.
   */
  return_url?: string;
  /**
   * string
   *
   * DEPRECATED. The URL where the customer is redirected after the customer cancels the payment.
   */
  cancel_url?: string;
  /**
   * object
   *
   * DEPRECATED. The customer and merchant payment preferences.
   */
  payment_method?: PaymentMethod;
  /**
   * object
   *
   * Provides additional details to process a payment using a payment_source that has been stored or is intended to be stored (also referred to as stored_credential or card-on-file).
   */
  stored_payment_source?: StoredPaymentSource;
}
// ...
export interface Bancontact {
  /**
   * string
   *
   * The name of the account holder associated with this payment method.
   */
  name: string;
  /**
   * string
   *
   * The two-character ISO 3166-1 country code.
   */
  country_code: string;
  experience_context: Pick<
    ExperienceContextPaypal,
    'brand_name' | 'shipping_preference' | 'locale' | 'return_url' | 'cancel_url'
  >;
}
// ...
export interface PaymentSource {
  /**
   * The payment card to use to fund a payment. Can be a credit or debit card.
   */
  card?: CardPaymentSource;
  /**
   * object
   *
   * The tokenized payment source to fund a payment.
   */
  token?: {
    /**
     * string
     *
     * The PayPal-generated ID for the token.
     *
     * required
     */
    id: string;
    /**
     * string
     *
     * The tokenization method that generated the ID.
     *
     * required
     */
    type: string;
  };
  /**
   * Indicates that PayPal Wallet is the payment source. Main use of this selection is to provide additional instructions associated with this choice like vaulting.
   */
  paypal?: PaypalPaymentSource;
  /**
   * object
   *
   * Bancontact is the most popular online payment in Belgium
   */
  bancontact?: Bancontact;
  /**
   * object
   *
   * BLIK is a mobile payment system, created by Polish Payment Standard in order to allow millions of users to pay in shops, payout cash in ATMs and make online purchases and payments.
   */
  blik?: Blik;
  /**
   * object
   *
   * The eps transfer is an online payment method developed by many Austrian banks.
   */
  eps?: Esp;
  /**
   * object
   *
   * Giropay is an Internet payment System in Germany, based on online banking.
   */
  giropay?: Giropay;
  /**
   * object
   *
   * The Dutch payment method iDEAL is an online payment method that enables consumers to pay online through their own bank.
   */
  ideal?: Ideal;
  /**
   * object
   *
   * MyBank is an e-authorisation solution which enables safe digital payments and identity authentication through a consumer’s own online banking portal or mobile application.
   */
  mybank?: Mybank;
  /**
   * object
   *
   * P24 (Przelewy24) is a secure and fast online bank transfer service linked to all the major banks in Poland.
   */
  p24?: P24;
  /**
   * object
   *
   * SOFORT Banking is a real-time bank transfer payment method that buyers use to transfer funds directly to merchants from their bank accounts.
   */
  sofort?: Sofort;
  /**
   * object
   *
   * Trustly is a payment method that allows customers to shop and pay from their bank account.
   */
  trustly?: Trustly;
  /**
   * object
   *
   * ApplePay payment source, allows buyer to pay using ApplePay, both on Web as well as on Native.
   */
  apple_pay?: ApplePay;
  /**
   * object
   *
   * Information needed to indicate that Venmo is being used to fund the payment.
   */
  venmo?: Venmo;
}

export interface CreateOrderRequest {
  /**
   * Array of objects [ 1 .. 10 ] items
   *
   * An array of purchase units. Each purchase unit establishes a contract between a payer and the payee. Each purchase unit represents either a full or partial order that the payer intends to purchase from the payee.
   *
   * required
   */
  purchase_units: PurchaseUnit[];
  /**
   *
   * string
   *
   * The intent to either capture payment immediately or authorize a payment for an order after order creation.
   *
   * required
   */
  intent: Intent;
  /**
   * object
   *
   * DEPRECATED. The customer is also known as the payer. The Payer object was intended to only be used with the payment_source.paypal object.
   * In order to make this design more clear, the details in the payer object are now available under payment_source.paypal. Please use payment_source.paypal.
   */
  payer?: Payer | undefined;

  payment_source?: PaymentSource | undefined;
  /**
   * object
   *
   * Customize the payer experience during the approval process for the payment with PayPal.
   */
  application_context?: ApplicationContext | undefined;
}